refactor(bhwi): isolate Jade common adapters
- move command and descriptor shaping to the common boundary
- map Jade recipients, transmits, responses, and errors externally
- keep the CBOR interpreter focused on Jade-native types
