+++ bhwi/src/common/adapters.rs
@@ -1,2 +1,3 @@
 mod coldcard;
+mod jade;
 mod ledger;
