+++ bhwi/src/jade/mod.rs
@@ -6,22 +6,15 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 
 use base64ct::{Base64, Encoding};
 use bitcoin::Network;
-use bitcoin::address::AddressType;
 use bitcoin::bip32::{DerivationPath, Fingerprint, Xpub};
-use bitcoin::hashes::{Hash, sha256};
 use bitcoin::psbt::Psbt;
 use bitcoin::secp256k1::ecdsa::Signature;
 use serde::Serialize;
 use serde::de::DeserializeOwned;
 
 use crate::Interpreter;
-use crate::common::{
-    Command, DisplayAddress, Error, Info, MultisigAddressType, MultisigDisplayAddress, Recipient,
-    Response, Transmit,
-};
 use crate::device::DeviceId;
 use crate::jade::api::GetInfoResponse;
-use crate::miniscript::descriptor::{DescriptorPublicKey, Wildcard};
 
 pub const JADE_NETWORK_MAINNET: &str = "mainnet";
 pub const JADE_NETWORK_TESTNET: &str = "testnet";
@@ -523,217 +516,15 @@ where
     }
 }
 
-impl TryFrom<Command> for JadeCommand {
-    type Error = Error;
-
-    fn try_from(cmd: Command) -> Result<Self, Self::Error> {
-        match cmd {
-            Command::Setup(..) => Err(Error::MissingCommandInfo("Setup not supported by Jade")),
-            Command::Wipe => Err(Error::MissingCommandInfo("Wipe not supported by Jade")),
-            Command::Restore(..) => Err(Error::MissingCommandInfo("Restore not supported by Jade")),
-            Command::TogglePassphrase => Err(Error::MissingCommandInfo(
-                "Toggle passphrase not supported by Jade",
-            )),
-            Command::Backup => Err(Error::MissingCommandInfo("Backup not supported by Jade")),
-            Command::Unlock { .. } => Ok(Self::Auth),
-            Command::GetMasterFingerprint => Ok(Self::GetMasterFingerprint),
-            Command::GetXpub { path, .. } => Ok(Self::GetXpub(path)),
-            Command::DisplayAddress(
-                DisplayAddress::ByDescriptor {
-                    index,
-                    change,
-                    descriptor_name,
-                    ..
-                },
-                _ctx,
-            ) => Ok(Self::GetReceiveAddress(ReceiveAddress::Descriptor {
-                index,
-                change,
-                descriptor_name,
-            })),
-            Command::DisplayAddress(
-                DisplayAddress::ByPath {
-                    path,
-                    address_format,
-                    ..
-                },
-                _,
-            ) => Ok(Self::GetReceiveAddress(ReceiveAddress::Path {
-                path,
-                variant: jade_path_variant(address_format)?,
-            })),
-            Command::DisplayAddress(DisplayAddress::ByMultisig(address), _) => {
-                jade_multisig_command(address)
-            }
-            Command::SignMessage { message, path } => Ok(Self::SignMessage { message, path }),
-            Command::GetVersion => Ok(Self::GetInfo),
-            Command::RegisterWallet { name, policy } => {
-                let (descriptor, keys) = crate::policy::extract_parts(&policy)
-                    .map_err(|err| Error::Serialization(err.to_string()))?;
-                // Jade's descriptor parser uses the explicit multipath spelling,
-                // while BIP-388 WalletPolicy display canonicalizes it to `/**`.
-                let descriptor = descriptor.replace("/**", "/<0;1>/*");
-                let datavalues = keys
-                    .iter()
-                    .enumerate()
-                    .map(|(index, key)| (format!("@{index}"), crate::policy::format_key_info(key)))
-                    .collect();
-                Ok(Self::RegisterDescriptor {
-                    descriptor_name: name,
-                    descriptor,
-                    datavalues,
-                })
-            }
-            Command::SignTx(psbt, _) => Ok(Self::SignPsbt { psbt }),
-        }
-    }
-}
-
-fn jade_multisig_command(address: MultisigDisplayAddress) -> Result<JadeCommand, Error> {
-    let variant = match address.address_type {
-        MultisigAddressType::Legacy => "sh(multi(k))",
-        MultisigAddressType::Wit => "wsh(multi(k))",
-        MultisigAddressType::ShWit => "sh(wsh(multi(k)))",
-    };
-    let mut signer_origins = Vec::with_capacity(address.keys.len());
-    let mut signers = Vec::with_capacity(address.keys.len());
-    let mut paths = Vec::with_capacity(address.keys.len());
-
-    for key in address.keys {
-        let DescriptorPublicKey::XPub(key) = key else {
-            return Err(Error::InvalidInput(
-                "Jade multisig display requires extended public keys".into(),
-            ));
-        };
-        if key.wildcard != Wildcard::None {
-            return Err(Error::InvalidInput(
-                "Jade multisig display requires concrete key derivation paths".into(),
-            ));
-        }
-        let (fingerprint, origin_path) = key.origin.ok_or_else(|| {
-            Error::InvalidInput("Jade multisig display requires key origin information".into())
-        })?;
-        let origin = origin_path.to_u32_vec();
-        signer_origins.push((fingerprint.to_bytes(), origin.clone()));
-        signers.push(api::MultisigSigner {
-            fingerprint: fingerprint.to_bytes().to_vec(),
-            derivation: origin,
-            xpub: key.xkey.to_string(),
-            path: Vec::new(),
-        });
-        paths.push(key.derivation_path.to_u32_vec());
-    }
-
-    signer_origins.sort();
-    let mut summary = format!("{variant}|{}|", address.threshold);
-    for (fingerprint, path) in signer_origins {
-        summary.push_str(&hex::encode(fingerprint));
-        summary.push('|');
-        summary.push_str(&format!("{path:?}"));
-        summary.push('|');
-    }
-    let digest = sha256::Hash::hash(summary.as_bytes());
-    let multisig_name = format!("hwi{}", &digest.to_string()[..12]);
-
-    Ok(JadeCommand::RegisterMultisig {
-        multisig_name,
-        descriptor: api::MultisigDescriptor {
-            variant: variant.to_string(),
-            sorted: address.sorted,
-            threshold: address.threshold,
-            signers,
-            master_blinding_key: None,
-        },
-        paths,
-    })
-}
-
-fn jade_path_variant(address_format: Option<AddressType>) -> Result<&'static str, Error> {
-    match address_format.unwrap_or(AddressType::P2wpkh) {
-        AddressType::P2pkh => Ok("pkh(k)"),
-        AddressType::P2sh => Ok("sh(wpkh(k))"),
-        AddressType::P2wpkh => Ok("wpkh(k)"),
-        AddressType::P2wsh | AddressType::P2tr => Err(Error::UnsupportedDisplayAddress(
-            "Jade does not support this path address format".into(),
-        )),
-        _ => Err(Error::UnsupportedDisplayAddress(
-            "Jade does not support this path address format".into(),
-        )),
-    }
-}
-
-impl From<JadeResponse> for Response {
-    fn from(res: JadeResponse) -> Response {
-        match res {
-            JadeResponse::TaskDone => Response::TaskDone,
-            JadeResponse::MasterFingerprint(fg) => Response::MasterFingerprint(fg),
-            JadeResponse::Xpub(xpub) => Response::Xpub(xpub),
-            JadeResponse::Signature(header, signature) => Response::Signature(header, signature),
-            JadeResponse::GetInfo(info) => Response::Info(Info {
-                version: info.jade_version.as_str().into(),
-                networks: info.jade_networks.into(),
-                firmware: None,
-                initialized: None,
-            }),
-            JadeResponse::Address(address) => Response::Address(address),
-            JadeResponse::RegisteredDescriptor => {
-                Response::WalletRegistration(crate::common::WalletRegistration::Complete {
-                    hmac: None,
-                })
-            }
-            JadeResponse::SignedPsbt(psbt) => Response::SignedPsbt(psbt),
-        }
-    }
-}
-
-impl From<JadeRecipient> for Recipient {
-    fn from(recipient: JadeRecipient) -> Recipient {
-        match recipient {
-            JadeRecipient::Device => Recipient::Device,
-            JadeRecipient::PinServer { url } => Recipient::PinServer { url },
-        }
-    }
-}
-
-impl From<JadeTransmit> for Transmit {
-    fn from(transmit: JadeTransmit) -> Transmit {
-        Transmit {
-            recipient: transmit.recipient.into(),
-            payload: transmit.payload,
-            encrypted: false,
-        }
-    }
-}
-
-impl From<JadeError> for Error {
-    fn from(error: JadeError) -> Error {
-        match error {
-            JadeError::Cbor => Error::Serialization("cbor".to_string()),
-            JadeError::NoErrorOrResult => Error::NoErrorOrResult,
-            JadeError::Rpc(api_error) => Error::Rpc(api_error.code, api_error.message),
-            JadeError::Serialization(s) => Error::Serialization(s),
-            JadeError::UnexpectedResult(msg) => Error::unexpected_result(
-                msg.clone().into_bytes(),
-                format!("jade unexpected result: {msg}"),
-            ),
-            JadeError::HandshakeRefused => Error::AuthenticationRefused,
-            JadeError::UnsupportedDisplayAddress => {
-                Error::UnsupportedDisplayAddress("unsupported display address on Jade".into())
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
 
+    use bitcoin::address::AddressType;
+
     use super::*;
     use crate::Interpreter;
-    use crate::common::{
-        Command, DisplayAddress, JadeInterpreter, MultisigAddressType, MultisigDisplayAddress,
-    };
-    use crate::miniscript::descriptor::DescriptorPublicKey;
+    use crate::common::{Command, DisplayAddress, JadeInterpreter, Response, WalletRegistration};
     use crate::miniscript::descriptor::WalletPolicy;
     use serde::Deserialize;
 
@@ -792,20 +583,6 @@ mod tests {
     }
 
     #[test]
-    fn path_display_rejects_unsupported_jade_address_format() {
-        let result = JadeCommand::try_from(Command::DisplayAddress(
-            DisplayAddress::ByPath {
-                path: "m/86'/1'/0'/0/0".parse().unwrap(),
-                display: true,
-                address_format: Some(AddressType::P2tr),
-            },
-            None,
-        ));
-
-        assert!(matches!(result, Err(Error::UnsupportedDisplayAddress(_))));
-    }
-
-    #[test]
     fn register_wallet_encodes_jade_descriptor_params() {
         let mut interpreter = JadeInterpreter::default().with_network(Network::Testnet);
         let transmit = interpreter
@@ -856,9 +633,7 @@ mod tests {
         assert!(interpreter.exchange(response).unwrap().is_none());
         assert!(matches!(
             interpreter.end().unwrap(),
-            Response::WalletRegistration(crate::common::WalletRegistration::Complete {
-                hmac: None
-            })
+            Response::WalletRegistration(WalletRegistration::Complete { hmac: None })
         ));
     }
 
@@ -889,44 +664,5 @@ mod tests {
                 .to_string()
                 .contains("register_descriptor returned false")
         );
-    }
-
-    #[test]
-    fn multisig_display_uses_upstream_hwi_registration_shape_and_name() {
-        let keys = [
-            "[f5acc2fd/48'/1'/0'/2']tpubDCbK3Ysvk8HjcF6mPyrgMu3KgLiaaP19RjKpNezd8GrbAbNg6v5BtWLaCt8FNm6QkLseopKLf5MNYQFtochDTKHdfgG6iqJ8cqnLNAwtXuP/0/7",
-            "[00000000/48'/1'/0'/2']tpubDDtb2WPYwEWw2WWDV7reLV348iJHw2HmhzvPysKKrJw3hYmvrd4jasyoioVPdKGQqjyaBMEvTn1HvHWDSVqQ6amyyxRZ5YjpPBBGjJ8yu8S/0/7",
-        ]
-        .map(|key| DescriptorPublicKey::from_str(key).unwrap())
-        .to_vec();
-
-        let command = jade_multisig_command(MultisigDisplayAddress {
-            threshold: 2,
-            address_type: MultisigAddressType::Wit,
-            sorted: true,
-            keys,
-        })
-        .unwrap();
-
-        let JadeCommand::RegisterMultisig {
-            multisig_name,
-            descriptor,
-            paths,
-        } = command
-        else {
-            panic!("expected multisig registration");
-        };
-        assert_eq!(multisig_name, "hwi78631c5c8b92");
-        assert_eq!(descriptor.variant, "wsh(multi(k))");
-        assert!(descriptor.sorted);
-        assert_eq!(descriptor.threshold, 2);
-        assert_eq!(descriptor.signers.len(), 2);
-        assert!(
-            descriptor
-                .signers
-                .iter()
-                .all(|signer| signer.path.is_empty())
-        );
-        assert_eq!(paths, vec![vec![0, 7], vec![0, 7]]);
     }
 }
